+++ examples/v2/next-pages-router/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @copilotkitnext/next-pages-router
 
+## 1.51.5-next.2
+
+### Patch Changes
+
+- @copilotkitnext/react@1.51.5-next.2
+
+## 1.51.5-next.1
+
+### Patch Changes
+
+- @copilotkitnext/react@1.51.5-next.1
+
 ## 1.51.5-next.0
 
 ### Patch Changes