+++ packages/v2/angular/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @copilotkitnext/angular
 
+## 1.51.5-next.2
+
+### Patch Changes
+
+- @copilotkitnext/core@1.51.5-next.2
+- @copilotkitnext/shared@1.51.5-next.2
+
+## 1.51.5-next.1
+
+### Patch Changes
+
+- @copilotkitnext/core@1.51.5-next.1
+- @copilotkitnext/shared@1.51.5-next.1
+
 ## 1.51.5-next.0
 
 ### Patch Changes