+++ packages/v2/core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @copilotkitnext/core
 
+## 1.51.5-next.2
+
+### Patch Changes
+
+- @copilotkitnext/shared@1.51.5-next.2
+
+## 1.51.5-next.1
+
+### Patch Changes
+
+- @copilotkitnext/shared@1.51.5-next.1
+
 ## 1.51.5-next.0
 
 ### Patch Changes